+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/amazon-bedrock
 
+## 3.0.59
+
+### Patch Changes
+
+- Updated dependencies [3a7a427]
+  - @ai-sdk/anthropic@2.0.47
+
+## 3.0.58
+
+### Patch Changes
+
+- Updated dependencies [54a0480]
+  - @ai-sdk/anthropic@2.0.46
+
 ## 3.0.57
 
 ### Patch Changes